services/ws_public_ingest/requirements.txt
websockets>=12.0
redis>=5.0.1
pyarrow>=16.0.0
fastparquet>=2024.2.0
prometheus-client>=0.20.0
pyyaml>=6.0.1
orjson>=3.9.15
